refactor: user.interface 를 User 정lass 로 수
