+++ backend/chatServer/src/user/user.dto.ts
@@ -1,0 +1,11 @@
+class User {
+  constructor(
+    public address: string,
+    public userAgent: string,
+    public entryTime: string,
+    public nickname: string,
+    public color: string
+  ) {}
+}
+
+export { User };
