--- backend/chatServer/src/user/user.interface.ts
@@ -1,9 +1,0 @@
-interface User {
-  address: string;
-  userAgent: string;
-  entryTime: string;
-  nickname: string;
-  color: string;
-}
-
-export { User };
